Delete validation when link add.
Because we don't know how to implement validation at the time of update, and it is meaningless to validate only at the time of registration. Also, it is not a big problem even if it is not validated, but it is a hindrance and causes stress if there is validation during registration.